authorization

+++ HRsystem.Api/Features/Scheduling/RemoteWorkdays/RemoteWorkdaysEndPoints.cs
@@ -4,6 +4,7 @@ using HRsystem.Api.Features.Scheduling.RemoteWorkdays.GetAllRemoteWorkdays;
 using HRsystem.Api.Features.Scheduling.RemoteWorkdays.GetRemoteWorkDaysById;
 using HRsystem.Api.Features.Scheduling.RemoteWorkdays.UpdateRemoteWorkDays;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 
 namespace HRsystem.Api.Features.Scheduling.RemoteWorkdays
 {
@@ -14,21 +15,21 @@ namespace HRsystem.Api.Features.Scheduling.RemoteWorkdays
             var group = app.MapGroup("/api/Scheduling/RemoteWorkDays").WithTags("RemoteWorkDays");
 
             // Get all
-            group.MapGet("/GetList", async (ISender mediator) =>
+            group.MapGet("/GetList", [Authorize] async (ISender mediator) =>
             {
                 var result = await mediator.Send(new GetAllRemoteWorkDaysQuery());
                 return Results.Ok(new { Success = true, Data = result });
             });
 
             // Create
-            group.MapPost("/Create", async (ISender mediator, CreateRemoteWorkDaysCommand command) =>
+            group.MapPost("/Create", [Authorize] async (ISender mediator, CreateRemoteWorkDaysCommand command) =>
             {
                 var id = await mediator.Send(command);
                 return Results.Ok(new { Success = true, Id = id });
             });
 
             // Update
-            group.MapPut("/UpdateOne/{id}", async (ISender mediator, int id, UpdateRemoteWorkDaysCommand command) =>
+            group.MapPut("/UpdateOne/{id}", [Authorize] async (ISender mediator, int id, UpdateRemoteWorkDaysCommand command) =>
             {
                 if (id != command.Id) return Results.BadRequest();
                 var success = await mediator.Send(command);
@@ -36,13 +37,13 @@ namespace HRsystem.Api.Features.Scheduling.RemoteWorkdays
             });
 
             // Delete
-            group.MapDelete("/DeleteOne/{id}", async (ISender mediator, int id) =>
+            group.MapDelete("/DeleteOne/{id}", [Authorize] async (ISender mediator, int id) =>
             {
                 var success = await mediator.Send(new DeleteRemoteWorkDaysCommand(id));
                 return success ? Results.Ok(new { Success = true }) : Results.NotFound();
             });
 
-            group.MapGet("/GetOne/{id}", async (int id, ISender mediator) =>
+            group.MapGet("/GetOne/{id}", [Authorize] async (int id, ISender mediator) =>
             {
                 var result = await mediator.Send(new GetRemoteWorkDaysByIdQuery(id));
                 return result == null
